StructuralCausalModels

Documentation goes here.

scm_path

StructuralCausalModels.scm_path — Method

scm_path

Relative path using the StatisticalRethinking src/ directory. Copied from DynamicHMCExamples.jl

Example to get access to the data subdirectory

scm_path("..", "data")

DAG

StructuralCausalModels.DAG — Type

DAG

Directed acyclic graph struct

Struct

DAG(
* `name::AbstractString`                    : Variables used to compute correlation
* `d::OrderedDict{Symbol, Vector{Symbol}}`  : DAG definition aas a Dict
* `a::NamedArray`                           : Adjacency matrix
* `e::NamedArray`                           : Edge matric
* `s::NamedArray`                           : Covariance matrix
* `df::DataFrame`                           : Variable observations
* `vars::Vector{Symbol}`                    : Names of variables
)

Part of API, exported.

StructuralCausalModels.DAG — Method

DAG

Directed acyclic graph constructor

DAG(name, d, df)

Method

DAG(
* `name::AbstractString`                    : Variables used to compute correlation
* `d::OrderedDict{Symbol, Vector{Symbol}}`  : DAG definition aas a Dict
* `df::DataFrame`                           : Variable observations
)

Returns

* `res::DAG`                                : Boolean result of test

Extended help

Example

Define and create a DAG

using StructuralCausalModels, CSV

df = CSV.read(scm_path("..", "data", "marks.csv");

d = OrderedDict(
  :mechanics => [:vectors, :algebra],
  :vectors => [:algebra],
  :analysis => [:algebra],
  :statistics => [:algebra, :analysis]
);

dag = DAG("marks", d, df);

Acknowledgements

Original author: Giovanni M. Marchetti

Translated to Julia: Rob J Goedman

License

The R package ggm is licensed under License: GPL-2.

The Julia translation is licenced under: MIT.

Part of API, exported.

d_separation

StructuralCausalModels.d_separation — Function

d_separation

d_separation(d, first, second)
d_separation(d, first, second, cond; debug)

Computes the d_separation between 2 sets of nodes conditioned on a third set.

Method

d_separation(
* `d::DAG`                             : DAG
* `first::Vector{Symbol}`              : First set
* `second::Vector{Symbol}`             : Second set
* `cond::Vector{Symbol}`               : Conditioning set
)

Returns

* `res::Bool`                          : Boolean result of test

Extended help

Example

d_separation between mechanics and statistics, conditioning on algebra

using StructuralCausalModels, CSV

df = CSV.read(scm_path("..", "data", "marks.csv");

d = OrderedDict(
  :mechanics => [:vectors, :algebra],
  :vectors => [:algebra],
  :analysis => [:algebra],
  :statistics => [:algebra, :analysis]
);

dag = DAG("marks", d, df);
d_separation(marks, [:statistics], [:mechanics], [:algebra]))

Acknowledgements

Original author: Giovanni M. Marchetti

Translated to Julia: Rob J Goedman

License

The R package ggm is licensed under License: GPL-2.

The Julia translation is licenced under: MIT.

Part of the API, exported.

##basis_set

StructuralCausalModels.basis_set — Method

basis_set

Determine basis_set

basis_set(dag)

Part of API, not exported.

##shipley_test

StructuralCausalModels.shipley_test — Method

shipley_test

shipley_test(d)

Test of all independencies implied by a given DAG

Computes a simultaneous test of all independence relationships implied by a given Gaussian model defined according to a directed acyclic graph, based on the sample covariance matrix.

The test statistic is C = -2 sum(ln(pj)) where pj are the p-values of tests of conditional independence in the basis set computed by basiSet(A). The p-values are independent uniform variables on (0,1) and the statistic has exactly a chi square distribution on 2k degrees of freedom where k is the number of elements of the basis set. Shipley (2002) calls this test Fisher's C test.

Method

shipley_test(;
* `d::Dag`                             : Directed acyclic graph
)

Returns

* `res::NamedTuple`                    : (ctest=..., dof=..., pval=...)

where:

ctest: Test statistic C dof: Degrees of freedom. pval: The P-value of the test, assuming a two-sided alternative.

Extended help

Example

Shipley_test for the mathematics marks data

using StructuralCausalModels, RData

objs = RData.load(scm_path("..", "data", "marks.rda");
marks_df = objs["marks"]

d = OrderedDict(
  :mechanics => [:vectors, :algebra],
  :vectors => [:algebra],
  :statistics => [:algebra, :analysis],
  :analysis => [:algebra]
);
dag = Dag(d; df=df)
shipley_test(dag)

See also

?Dag
?basis_set
?pcor_test

Acknowledgements

Original author: Giovanni M. Marchetti

Translated to Julia: Rob J Goedman

References

Shipley, B. (2000). A new inferential test for path models based on directed acyclic graphs. Structural Equation Modeling, 7(2), 206–218.

Licence

The R package ggm is licensed under License: GPL-2.

The Julia translation is licenced under: MIT.

Exported as part of the api

##pcor

StructuralCausalModels.pcor — Method

pcor

pcor(u, S)

Computes the partial correlation between two variables given a set of other variables.

Method

pcor(;
* `u::Vector{Int}`                     : Variables used to compute correlation
* `S::Matrix`                          : Sample covariance matrix
)

where:

u[1], u[2]: Variables used to compute correlation between, remaining indices are the conditioning set

Returns

* `res::Float64`                       : Correlation between u[1] and u[2]

Extended help

Example

Correlation between vectors and algebra, conditioning on analysis and statistics

using StructuralCausalModels, CSV

df = CSV.read(scm_path("..", "data", "marks.csv");
S = cov(Array(df))

u = [2, 3, 4, 5]
pcor(u, S)
u = [:vectors, :algebra, :statistics, :analysis]

Acknowledgements

Original author: Giovanni M. Marchetti

Translated to Julia: Rob J Goedman

License

The R package ggm is licensed under License: GPL-2.

The Julia translation is licenced under: MIT.

Part of the api, not exported.