scm_path

StructuralCausalModels.scm_path — Method

scm_path

Relative path using the StructuralCausalModels.jl src/ directory.

Example to get access to the data subdirectory

scm_path("..", "data")

DAG

StructuralCausalModels.DAG — Type

DAG

Directed acyclic graph struct

Struct

DAG(
* `name::AbstractString`                    : Name for the DAG object
* `d::OrderedDictOrNothing`                 : DAG definition as an OrderedDict
* `a::NamedArrayOrNothing`                  : Adjacency matrix
* `e::NamedArrayOrNothing`                  : Edge matrix
* `s::NamedArrayOrNothing`                  : Covariance matrix
* `df::DataFrameOrNothing`                  : Variable observations
* `vars::Vector{Symbol}`                    : Names of variables in DAG
)

Part of API, exported.

StructuralCausalModels.DAG — Method

DAG

Directed acyclic graph constructor

DAG(name, d, df)

Required arguments

* `name::AbstractString`               : Name for the DAG object
* `d`                                  : DAG definition as an
                                           OrderedDict (see extended help)
                                           AbstractString (as in ggm or dagitty)
                                           AdjacencyMatrix

Optional positional argument

* `df::DataFrame`                      : DataFrame with observations

Returns

* `dag::DAG`                           : Boolean result of test

Extended help

In the definition of the OrderedDict, read => as ~ in regression models or <- in causal models, e.g.:

d = OrderedDict(
  :u => [:x, :v],
  :s1 => [:u],
  :w => [:v, :y],
  :s2 => [:w]
);
dag = DAG("my_name", d)

Coming from R's dagitty:

amat <- dagitty("dag { {X V} -> U; S1 <- U; {Y V} -> W; S2 <- W}”)

dag = DAG("my_name", "dag { {X V} -> U; S1 <- U; {Y V} -> W; S2 <- W}”)
display(dag.a) # Show the adjacency_matrix

Coming from R's ggm:

amat <- DAG(U~X+V, S1~U, W~V+Y, S2~W, order=FALSE)

dag = DAG("my_name", "DAG(U~X+V, S1~U, W~V+Y, S2~W”)
display(dag.a) # Show the adjacency_matrix

Acknowledgements

Original author: Giovanni M. Marchetti

Translated to Julia: Rob J Goedman

License

The R package ggm is licensed under License: GPL-2.

The Julia translation is licenced under: MIT.

Part of API, exported.

Missing docstring.

Missing docstring for DAG(name::AbstractString, d::OrderedDict). Check Documenter's build log for details.

Missing docstring.

Missing docstring for DAG(name::AbstractString, str::AbstractString, df::DataFrame). Check Documenter's build log for details.

Missing docstring.

Missing docstring for DAG(name::AbstractString, str::AbstractString). Check Documenter's build log for details.

Missing docstring.

Missing docstring for DAG(name::AbstractString, a::NamedArray, df::DataFrame). Check Documenter's build log for details.

Missing docstring.

Missing docstring for DAG(name::AbstractString, a::NamedArray). Check Documenter's build log for details.

d_separation

StructuralCausalModels.d_separation — Function

d_separation

d_separation(d, first, second)
d_separation(d, first, second, cond; debug)

Computes the d_separation between 2 sets of nodes conditioned on a third set.

Required arguments

d_separation(
* `d::DAG`                             : DAG
* `first::Vector{Symbol}`              : First set
* `second::Vector{Symbol}`             : Second set
)

Optional arguments

* `cond::Vector{Symbol}`               : Conditioning set
* `debug=false`                        : Trace execution

Returns

* `res::Bool`                          : Boolean result of test

Extended help

Example

d_separation between mechanics and statistics, conditioning on algebra

using StructuralCausalModels, CSV

df = CSV.read(scm_path("..", "data", "marks.csv");

d = OrderedDict(
  :mechanics => [:vectors, :algebra],
  :vectors => [:algebra],
  :analysis => [:algebra],
  :statistics => [:algebra, :analysis]
);

dag = DAG("marks", d, df);
d_separation(marks, [:statistics], [:mechanics], [:algebra]))

Acknowledgements

Original author: Giovanni M. Marchetti

Translated to Julia: Rob J Goedman

License

The R package ggm is licensed under License: GPL-2.

The Julia translation is licenced under: MIT.

Part of the API, exported.

shipley_test

StructuralCausalModels.shipley_test — Method

shipley_test

shipley_test(d)

Test of all independencies implied by a given DAG

Computes a simultaneous test of all independence relationships implied by a given Gaussian model defined according to a directed acyclic graph, based on the sample covariance matrix.

The test statistic is C = -2 sum(ln(pj)) where pj are the p-values of tests of conditional independence in the basis set computed by basiSet(A). The p-values are independent uniform variables on (0,1) and the statistic has exactly a chi square distribution on 2k degrees of freedom where k is the number of elements of the basis set. Shipley (2002) calls this test Fisher's C test.

Method

shipley_test(;
* `d::Dag`                             : Directed acyclic graph
)

Returns

* `res::NamedTuple`                    : (ctest=..., dof=..., pval=...)

where:

ctest: Test statistic C dof: Degrees of freedom. pval: The P-value of the test, assuming a two-sided alternative.

Extended help

Example

Shipley_test for the mathematics marks data

using StructuralCausalModels, RData

objs = RData.load(scm_path("..", "data", "marks.rda");
marks_df = objs["marks"]

d = OrderedDict(
  :mechanics => [:vectors, :algebra],
  :vectors => [:algebra],
  :statistics => [:algebra, :analysis],
  :analysis => [:algebra]
);
dag = Dag(d; df=df)
shipley_test(dag)

See also

?Dag
?basis_set
?pcor_test

Acknowledgements

Original author: Giovanni M. Marchetti

Translated to Julia: Rob J Goedman

References

Shipley, B. (2000). A new inferential test for path models based on directed acyclic graphs. Structural Equation Modeling, 7(2), 206–218.

Licence

The R package ggm is licensed under License: GPL-2.

The Julia translation is licenced under: MIT.

Part of the api, exported.

basis_set

StructuralCausalModels.basis_set — Method

basis_set

Determine basis_set

basis_set(dag)

Part of API, exported.

ancestral_graph

Missing docstring.

Missing docstring for ancestral_graph(d::DAG, m::Vector{Symbol}, c::Vector{Symbol}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for ancestral_graph(a::NamedArray{Int, 2}, m::Vector{Symbol}, c::Vector{Symbol}). Check Documenter's build log for details.

adjustment_sets

Missing docstring.

Missing docstring for adjustment_sets(d::DAG, f::Symbol, l::Symbol, u::Vector{Symbol}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for adjustment_sets(d::DAG, f::Symbol, l::Symbol). Check Documenter's build log for details.

paths

Missing docstring.

Missing docstring for all_paths(d::DAG, f::Symbol, l::Symbol). Check Documenter's build log for details.

Missing docstring.

Missing docstring for backdoor_paths(d::DAG, paths::Vector{Vector{Symbol}}, f::Symbol). Check Documenter's build log for details.

Missing docstring.

Missing docstring for check_open(d::DAG, path::Vector{Symbol}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for open_paths(d::DAG, paths::Vector{Vector{Symbol}}). Check Documenter's build log for details.

Internals

StructuralCausalModels.adjacency_matrix — Method

adjacency_matrix

adjacency_matrix(d)

Internal

StructuralCausalModels.adjacency_matrix — Method

adjacency_matrix

adjacency_matrix(e)

Internal

StructuralCausalModels.adjacency_matrix_to_dict — Method

adjacency_matrix to OrderedDict

adjacency_matrix_to_dict(ea)

Internal

StructuralCausalModels.ancester_graph — Method

ancestor_graph

ancester_graph(e)

Internal

Missing docstring.

Missing docstring for all_edges(a::NamedArray). Check Documenter's build log for details.

Missing docstring.

Missing docstring for blocking_sets(asets::Array{Array{Symbol,1},1}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for dag_show(io::IO, d::DAG). Check Documenter's build log for details.

StructuralCausalModels.dag_vars — Method

dag_vars

dag_vars(d)

Internal

StructuralCausalModels.edge_matrix — Method

edge_matrix

edge_matrix(d)

Internal

StructuralCausalModels.edge_matrix — Function

edge_matrix

edge_matrix(a)
edge_matrix(a, inv)

Internal

Missing docstring.

Missing docstring for forward_path(d::DAG, path). Check Documenter's build log for details.

Missing docstring.

Missing docstring for handle_rhs!(vars::Vector{Symbol}, rhs::SymbolList). Check Documenter's build log for details.

StructuralCausalModels.indicator_matrix — Method

indicator_matrix

indicator_matrix(e)

Internal

StructuralCausalModels.induced_covariance_graph — Method

inducedcovariancegraph

induced_covariance_graph(d, sel, cond; debug)

Internal

Missing docstring.

Missing docstring for node_edges(p::Path, s::Symbol, l::Symbol). Check Documenter's build log for details.

StructuralCausalModels.pcor — Method

pcor

pcor(u, S)

Computes the partial correlation between two variables given a set of other variables.

Method

pcor(;
* `u::Vector{Symbol}`                  : Variables used to compute correlation
* `S::Matrix`                          : Sample covariance matrix
)

where:

u[1], u[2]: Variables used to compute correlation between, remaining variables are the conditioning set

Returns

* `res::Float64`                       : Correlation between u[1] and u[2]

Extended help

Example

Correlation between vectors and algebra, conditioning on analysis and statistics

using StructuralCausalModels, CSV

df = CSV.read(scm_path("..", "data", "marks.csv");
S = cov(Array(df))

u = [2, 3, 4, 5]
pcor(u, S)
u = [:vectors, :algebra, :statistics, :analysis]

Acknowledgements

Original author: Giovanni M. Marchetti

Translated to Julia: Rob J Goedman

License

The R package ggm is licensed under License: GPL-2.

The Julia translation is licenced under: MIT.

Part of the api, not exported.

StructuralCausalModels.topological_order — Method

topological_order

topological_order(a)

Internal

StructuralCausalModels.topological_sort — Method

topological_sort

topological_sort(a)

Internal

StructuralCausalModels.transitive_closure — Method

transitive_closure

transitive_closure(a)

Internal

StructuralCausalModels.set_dag_df! — Method

setdagdf!

Set or update Dataframe associated to DAG

set_dag_df!(d, df; force)

Required arguments

* `d::DAG`                                  : Previously defined DAG object 
* `df::DataFrameOrNothing`                  : DataFrame associated with DAG
)

Optional arguments

* `force=false`                             : Force assignment of df 
)

The force = true option can be used if the DAG involves unobserved nodes.

StructuralCausalModels.set_dag_cov_matrix! — Method

setdagcov_matrix!

Set or update the covariance matrix associated to DAG

set_dag_cov_matrix!(d, cm; force)

Required arguments

* `d::DAG`                                  : Previously defined DAG object 
* `cm::NamedArrayOrNothing`                 : Covariance matrix in NamedArray format
)

Optional arguments

* `force=false`                             : Force assignment of df 
)

The force = true option can be used if the DAG involves unobserved nodes.

StructuralCausalModels.undirected_matrix — Method

undirected_matrix

undirected_matrix(d)

Internal